code cleanup
also I did something that changed how fast the temp updates. idk if it was in the max31855 library or... the goal was to get the pid algo better data, but it had the effect of making the temp change rapidly when the value is close to the next number. ie 29.9 30.0 29.9 etc. doesn't look good. i'd rather slow it down than have it 'flicker' like this